Rename example.py to example_app.py(#56)
Rename example.py to example_app.py and update references throughout the project
